Updated data_concatenation.ipynb to minimally process the raw data
And backed up the original file.
Made this change so that I have a standard minimally processed but concatenated ITP raw data file that I can use in multiple pipelines moving forward. This is because there are a few special cases of treatments that are going to be harder to deal with, like stopping a treatment after 3 months, or combining two drugs. In my first pipeline, I'll remove these for simplicity to get a working app up asap. Then I'll deal with those cases later - which is important because subsequent ITP cohorts also do things like use combinations, so I need to deal with those to apply this app to more recent data.